fix(og): move card dimensions to dep-free module, keep resvg out of client bundle
